EVM-IT: Use testkit::fevm
