+++ testing/integration/src/lib.rs
@@ -7,6 +7,4 @@ pub mod error;
 #[cfg(feature = "smt")]
 pub mod smt;
 pub mod tester;
-// TODO: Should come from https://github.com/filecoin-project/ref-fvm/pull/1493
-pub mod fevm;
 pub mod testkit;
