+++ testing/integration/tests/fevm_features/self_destruct.rs
@@ -3,7 +3,7 @@ use std::str::FromStr;
 use cucumber::gherkin::Step;
 use cucumber::{given, then, when, World};
 use ethers::types::H160;
-use fvm_integration_tests::fevm::EthAddress;
+use fvm_integration_tests::testkit::fevm::EthAddress;
 use fvm_shared::address::Address;
 
 use crate::common::*;
